Applied patch for character controller by bjorn.wesen. Fixes Issue 644.
Fixed some warnings in double-precision compile

+++ src/BulletMultiThreaded/vectormath2bullet.h
@@ -37,7 +37,7 @@
 
 inline Vectormath::Aos::Vector3	getVmVector3(const btVector3& bulletVec)
 {
-	return Vectormath::Aos::Vector3(bulletVec.getX(),bulletVec.getY(),bulletVec.getZ());
+	return Vectormath::Aos::Vector3((float)bulletVec.getX(),(float)bulletVec.getY(),(float)bulletVec.getZ());
 }
 
 inline btVector3 getBtVector3(const Vectormath::Aos::Vector3& vmVec)
@@ -51,7 +51,7 @@ inline btVector3 getBtVector3(const Vectormath::Aos::Point3& vmVec)
 
 inline Vectormath::Aos::Quat	getVmQuat(const btQuaternion& bulletQuat)
 {
-	Vectormath::Aos::Quat vmQuat(bulletQuat.getX(),bulletQuat.getY(),bulletQuat.getZ(),bulletQuat.getW());
+	Vectormath::Aos::Quat vmQuat((float)bulletQuat.getX(),(float)bulletQuat.getY(),(float)bulletQuat.getZ(),(float)bulletQuat.getW());
 	return vmQuat;
 }
